config: Move the config in a single configuration file
