+++ fedimg/uploader.py
@@ -29,8 +29,9 @@ import logging
 from fedimg.config import ACTIVE_SERVICES
 from fedimg.services.ec2.ec2initiate import main as ec2main
 from fedimg.services.ec2.ec2copy import main as ec2copy
-from fedimg.services.ec2.config import AWS_ACCESS_ID, AWS_SECRET_KEY
-from fedimg.services.ec2.config import AWS_BASE_REGION, AWS_REGIONS
+from fedimg.config import AWS_ACCESS_ID, AWS_SECRET_KEY
+from fedimg.config import AWS_BASE_REGION, AWS_REGIONS
+
 
 LOG = logging.getLogger(__name__)
